[eve MAINLINE] Automatic VCXPROJ submission from 889
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=3581758
!nobuild

Source Changes
----------------
[Eve Engine] 1389807 by olafurth (??lafur ????r Gunnarsson)
https://jira.ccpgames.com/browse/GFX-4637
Fixing unittests (that were passing on my machine but failing on the autobuilder)
----------------
[Eve Engine] 1389786 by olafurth (??lafur ????r Gunnarsson)
https://jira.ccpgames.com/browse/GFX-4461
Adding support for banner lights.
The banners will create a point light with an inner radius which should control how fuzzy the light should be
Review: http://codecollaborator/ui#review:id=16671
----------------
[Eve Engine] 1389785 by olafurth (??lafur ????r Gunnarsson)
https://jira.ccpgames.com/browse/GFX-4637
Adding the following methods to HostBitmap:
- GetPixelColor
- GetAverageColor
Added tests for all of the methods which should be working fine.
Review: http://codecollaborator/ui#review:id=16670
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/": change = 1389846]