[eve MAINLINE] Automatic VCXPROJ submission from 2349
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=9073607
!nobuild

Source Changes
----------------
[Eve Engine] 1665979 by filipp (Filipp Pavlov)
Defect: Missing support for DDS "metadata" in imagetools
https://jira.ccpgames.com/browse/EVEDEV-9423
Added support for text metadata to some image formats in ImageIO; exposed metadata for bitmaps in ImageTools.
This text metadata a list of key value strings. Formats supported:
- PNG - with text chunks
- TGA - with version 2 "developer area"
- DDS - with a hacky appending stuff past pixel data (similar to TGA)
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/": change = 1666156]